Just some small changes.

+++ List_src/List.cpp
@@ -28,7 +28,7 @@ int main(int argc, char** argv) {
 		
 		int n_inserts, n_removes, n_lookups;
 		fscanf(file, "%d%d%d%d", &option, &n_inserts, &n_removes, &n_lookups);
-		int balance_freq = floor(sqrt(n_inserts));
+		int balance_freq =  (int) floor(sqrt(n_inserts));
 		for (int i = 0; i < n_inserts; i++) {
 			fscanf(file, "%d", &v);
 			l->push_back(v);
@@ -45,6 +45,7 @@ int main(int argc, char** argv) {
 			fscanf(file, "%d", &v);
 			auto it = std::find(l->begin(), l->end(), v);
 		}
+
 		std::chrono::steady_clock::time_point end= std::chrono::steady_clock::now();
 		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(end - begin).count() << " milliseconds"<< std::endl;
 	}
